+++ nominal/core/__init__.py
@@ -1,19 +1,17 @@
 from __future__ import annotations
 
-import shutil
 import time
 import urllib.parse
 from dataclasses import dataclass, field
-from datetime import datetime, timedelta
+from datetime import timedelta
 from io import TextIOBase
 from pathlib import Path
 from types import MappingProxyType
-from typing import BinaryIO, Iterable, Mapping, Sequence, cast
+from typing import BinaryIO, Iterable, Mapping, Sequence
 
 from typing_extensions import Self
 
 from .._api.combined import (
-    attachments_api,
     datasource,
     datasource_logset_api,
     ingest_api,
@@ -25,6 +23,7 @@ from .._multipart import put_multipart_upload
 from .._utils import FileType, FileTypes, update_dataclass
 from ..exceptions import NominalIngestError, NominalIngestFailed, NominalIngestMultiError
 from ..ts import IntegralNanosecondsUTC, LogTimestampType, _AnyTimestampType, _SecondsNanos, _to_typed_timestamp_type
+from .attachment import Attachment
 from .client import NominalClient
 from .run import Run
 
@@ -230,76 +229,6 @@ class Log:
         if log.body.basic is None:
             raise RuntimeError(f"unhandled log body type: expected 'basic' but got {log.body.type!r}")
         return cls(timestamp=_SecondsNanos.from_api(log.time).to_nanoseconds(), body=log.body.basic.message)
-
-
-@dataclass(frozen=True)
-class Attachment:
-    rid: str
-    name: str
-    description: str
-    properties: Mapping[str, str]
-    labels: Sequence[str]
-    _client: NominalClient = field(repr=False)
-
-    def update(
-        self,
-        *,
-        name: str | None = None,
-        description: str | None = None,
-        properties: Mapping[str, str] | None = None,
-        labels: Sequence[str] | None = None,
-    ) -> Self:
-        """Replace attachment metadata.
-        Updates the current instance, and returns it.
-
-        Only the metadata passed in will be replaced, the rest will remain untouched.
-
-        Note: This replaces the metadata rather than appending it. To append to labels or properties, merge them before
-        calling this method. E.g.:
-
-            new_labels = ["new-label-a", "new-label-b", *attachment.labels]
-            attachment = attachment.update(labels=new_labels)
-        """
-        request = attachments_api.UpdateAttachmentRequest(
-            description=description,
-            labels=None if labels is None else list(labels),
-            properties=None if properties is None else dict(properties),
-            title=name,
-        )
-        response = self._client._attachment_client.update(self._client._auth_header, request, self.rid)
-        attachment = self.__class__._from_conjure(self._client, response)
-        update_dataclass(self, attachment, fields=self.__dataclass_fields__)
-        return self
-
-    def get_contents(self) -> BinaryIO:
-        """Retrieve the contents of this attachment.
-        Returns a file-like object in binary mode for reading.
-        """
-        response = self._client._attachment_client.get_content(self._client._auth_header, self.rid)
-        # note: the response is the same as the requests.Response.raw field, with stream=True on the request;
-        # this acts like a file-like object in binary-mode.
-        return cast(BinaryIO, response)
-
-    def write(self, path: Path, mkdir: bool = True) -> None:
-        """Write an attachment to the filesystem.
-
-        `path` should be the path you want to save to, i.e. a file, not a directory.
-        """
-        if mkdir:
-            path.parent.mkdir(exist_ok=True, parents=True)
-        with open(path, "wb") as wf:
-            shutil.copyfileobj(self.get_contents(), wf)
-
-    @classmethod
-    def _from_conjure(cls, client: NominalClient, attachment: attachments_api.Attachment) -> Self:
-        return cls(
-            rid=attachment.rid,
-            name=attachment.title,
-            description=attachment.description,
-            properties=MappingProxyType(attachment.properties),
-            labels=tuple(attachment.labels),
-            _client=client,
-        )
 
 
 @dataclass(frozen=True)
